requirements.txt
streamlit
pandas
plotly
prophet
requests
python-dotenv
